+++ .prettierignore
@@ -13,7 +13,7 @@ packages/graphql-yoga/src/landing-page-html.ts
 packages/render-graphiql/src/graphiql.ts
 .changeset/
 examples/apollo-federation-compatibility/src/resolvers-types.ts
-examples/tutorial/**/*.generated.*
+examples/hackernews/**/*.generated.*
 out/
 
 # Since prettier doesn't support MDX2 he breaks formatting for the following
