+++ examples/hackernews/.gitignore
@@ -3,3 +3,5 @@ node_modules
 .env
 prisma/migrations/migration_lock.toml
 prisma/*.db*
+# Ignore codegen auto-generated files
+*.generated.*
