+++ examples/hackernews/codegen.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from '@eddeee888/gcg-typescript-resolver-files';
+import type { CodegenConfig } from '@graphql-codegen/cli';
+
+const config: CodegenConfig = {
+  schema: 'src/schema/**/schema.graphql',
+  generates: {
+    'src/schema': defineConfig({
+      resolverGeneration: 'minimal',
+      typesPluginsConfig: {
+        contextType: '../context#GraphQLContext',
+      },
+    }),
+  },
+  hooks: { afterAllFileWrite: ['prettier --write'] },
+};
+export default config;
