+++ examples/hackernews/src/schema/base/resolvers/Query/info.ts
@@ -1,0 +1,5 @@
+import type { QueryResolvers } from '../../../types.generated';
+
+export const info: NonNullable<QueryResolvers['info']> = async (_parent, _arg, _ctx) => {
+  return `This is the API of a Hackernews Clone`;
+};
