+++ examples/hackernews/src/schema/base/resolvers/Query/link.ts
@@ -1,0 +1,7 @@
+import type { QueryResolvers } from '../../../types.generated';
+
+export const link: NonNullable<QueryResolvers['link']> = async (_parent, args, context) => {
+  return context.prisma.link.findUnique({
+    where: { id: parseInt(args.id) },
+  });
+};
