+++ examples/hackernews/src/schema/base/schema.graphql
@@ -1,0 +1,24 @@
+type Query {
+  info: String!
+  feed(filterNeedle: String, skip: Int, take: Int): [Link!]!
+  comment(id: ID!): Comment
+  link(id: ID!): Link
+}
+
+type Mutation {
+  postLink(url: String!, description: String!): Link!
+  postCommentOnLink(linkId: ID!, body: String!): Comment!
+}
+
+type Link {
+  id: ID!
+  description: String!
+  url: String!
+  comments: [Comment!]!
+}
+
+type Comment {
+  id: ID!
+  body: String!
+  link: Link
+}
