+++ examples/hackernews/src/schema/base/schema.mappers.ts
@@ -1,0 +1,4 @@
+import type { Comment, Link } from '@prisma/client';
+
+export type LinkMapper = Link;
+export type CommentMapper = Comment;
